% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BothHemisphere_met_age.R
\name{BothHemisphere_met_age}
\alias{BothHemisphere_met_age}
\title{Model two regions}
\usage{
BothHemisphere_met_age(d, region1, region2, metabolite, CRLB,
  saveplot = F)
}
\arguments{
\item{d}{dataframe with region, metabolite, and crlb columns}

\item{region1}{numeric region to include}

\item{region2}{numeric region to include}

\item{metabolite}{unquoted column name for y}

\item{CRLB}{unquoted column name for thres cutoff}

\item{saveplot}{boolean save or not}
}
\description{
`BothHemisphere_met_age` returns plot, model summary, and row count
}
\details{

}
