% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{fit_and_data}
\alias{fit_and_data}
\title{fit and filter data}
\usage{
fit_and_data(d, regions, mtbl, thres, crlb_thres = 20)
}
\arguments{
\item{d}{datafraem with region, metabolite, and crlb columns}

\item{regions}{regions to include, like c(1,2,3)}
}
\value{
list with d (cleaned data w/concentration and metabolite column names) and f (fitted data)
}
\description{
fitted and flitered data with common column names
}
