% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsi_fitdf.R
\name{mrsi_bestmodel}
\alias{mrsi_bestmodel}
\title{find best MRSI data model fit}
\usage{
mrsi_bestmodel(d, metabolite, nuissance = "GMrat")
}
\arguments{
\item{d}{(cleaned) dataframe with ld8, roi+label, metabolite, and CRLB columns}

\item{metabolite}{quoted column name (x value)}

\item{nuissance}{formula string for additional regressors. defaults to "GMrat"}
}
\description{
pick lowest AIC model among lin, inv, and quad w/rand effects if more than one region
}
