% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsi_fitdf.R
\name{mrsi_clean}
\alias{mrsi_clean}
\title{CRLB clean
add ages, remove bad crlb}
\usage{
mrsi_clean(
  d,
  regions,
  CRLB,
  nona = c("GMrat", CRLB),
  crlb_thres = 20,
  mesg = F
)
}
\arguments{
\item{d}{dataframe with ld8, region, metabolite, and CRLB columns}

\item{regions}{numeric regions vector to include}

\item{CRLB}{quoted column name for thresholding}

\item{crlb_thres}{where to discard SD values (default 20)}

\item{mesg}{boolean if we should report how many are kept/removed (default F)}
}
\description{
CRLB clean
add ages, remove bad crlb
}
