% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsi_fitdf.R
\name{mrsi_fitdf}
\alias{mrsi_fitdf}
\title{get estimates from the best model}
\usage{
mrsi_fitdf(m)
}
\arguments{
\item{m}{model generated by [mrsi_bestmodel()]}
}
\description{
returns estimate dataframe for plotting.
includes p and beta columns (same value repeated in ea. row)
}
