% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{mrsi_plot}
\alias{mrsi_plot}
\title{Plot best model fit}
\usage{
mrsi_plot(d, regions, metabolite, CRLB, saveplot = F)
}
\arguments{
\item{d}{dataframe with region, metabolite, and crlb columns}

\item{regions}{numeric region(s) to include}

\item{metabolite}{unquoted column name for y}

\item{CRLB}{unquoted column name for thres cutoff}

\item{saveplot}{boolean save or not}
}
\description{
return a plot of fits
}
\details{

}
