% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{mrsi_plot_many}
\alias{mrsi_plot_many}
\title{Plot best model fit over raw data for multiple metabolites}
\usage{
mrsi_plot_many(d, regions, mtbl_thres, saveplot = F)
}
\arguments{
\item{d}{dataframe with region, metabolite, and crlb columns}

\item{regions}{numeric region(s) to include}

\item{mtbl_thres}{list of quoted metabolite and thres column names
list(Glu.Cr="Glu.SD"), GABA.Cr="GABA.SD"))}

\item{saveplot}{boolean save or not}
}
\description{
return a plot of fits and raw
}
