% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsi_fitdf.R
\name{p_or_t}
\alias{p_or_t}
\title{p-value or t-value
returns whichever model supports}
\usage{
p_or_t(cfs, aslist = F)
}
\arguments{
\item{cfs}{single coefficent row from model summary.}

\item{aslist}{boolean to return list instead of string}
}
\description{
p-value or t-value
returns whichever model supports
}
