% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsi_fitdf.R
\name{mrsi_bestmodel}
\alias{mrsi_bestmodel}
\title{find best MRSI data model fit}
\usage{
mrsi_bestmodel(d, regions, metabolite, crlb_thres = 20, CRLB = NULL)
}
\arguments{
\item{d}{(cleaned) dataframe with ld8, region, metabolite, and CRLB columns}

\item{regions}{numeric regions vector to include}

\item{metabolite}{quoted column name (x value)}

\item{CRLB}{quoted column name for thresholding (x value)}
}
\description{
pick lowest AIC model among lin, inv, and quad w/rand effects if more than one region
}
